refactor: global state 확장성 고려한 리팩토링
