+++ src/components/backdrop/sidebar/app-sidebar.tsx
@@ -21,11 +21,6 @@ import { NavWorkSpace } from "./navWorkspace";
 import { useDataStore } from "@/src/stores/storeProvider";
 
 const data = {
-  user: {
-    name: "무제",
-    email: "untitled_01@gmail.com",
-    avatar: "https://github.com/shadcn.png",
-  },
   workspaces: [
     {
       name: "워크스페이스 1",
@@ -118,7 +113,7 @@ export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
         <NavPages label="페이지" pages={data.pages} />
       </SidebarContent>
       <SidebarFooter>
-        <NavUser user={globalUser} />
+        <NavUser user={globalUser.data} />
       </SidebarFooter>
       {/* <SidebarRail /> */}
     </Sidebar>
